SOC-1473 | Controller to component WiP
