+++ front/scripts/main/components/discussion-forum.js
@@ -1,0 +1,8 @@
+export default Ember.Component.extend(
+	{
+
+		actions: {
+
+		},
+	}
+);
