+++ front/scripts/main/components/forum-wrapper.js
@@ -1,7 +1,4 @@
-import DiscussionUpvoteActionSendMixin from '../mixins/discussion-upvote-action-send';
-
 export default Ember.Component.extend(
-	DiscussionUpvoteActionSendMixin,
 	{
 		classNames: ['forum-wrapper', 'discussion', 'forum'],
 		postsDisplayed: 0,
@@ -24,32 +21,11 @@ export default Ember.Component.extend(
 
 		actions: {
 			/**
-			 * @param {number} postId
-			 * @param {bool} openInNewTab
-			 * @returns {void}
-			 */
-			goToPost(postId, openInNewTab = false) {
-				this.sendAction('goToPost', postId, openInNewTab);
-			},
-
-			/**
 			 * @param {string} sortBy
 			 * @returns {void}
 			 */
 			setSortBy(sortBy) {
 				this.sendAction('setSortBy', sortBy);
-			},
-
-			/**
-			 * @param {any} postData
-			 * @returns {*}
-			 */
-			create(postData) {
-				return this.sendAction('create', postData);
-			},
-
-			toggleEditor(active) {
-				this.sendAction('toggleEditor', active);
 			}
 		},
