+++ front/scripts/main/components/post-detail.js
@@ -1,8 +1,6 @@
-import DiscussionUpvoteActionSendMixin from '../mixins/discussion-upvote-action-send';
 import DiscussionParsedContentMixin from '../mixins/discussion-parsed-content';
 
 export default Ember.Component.extend(
-	DiscussionUpvoteActionSendMixin,
 	DiscussionParsedContentMixin,
 	{
 		classNames: ['post-detail'],
@@ -51,9 +49,9 @@ export default Ember.Component.extend(
 			goToPost(postId, event) {
 				event.preventDefault();
 				if (event.ctrlKey || event.metaKey) {
-					this.sendAction('goToPost', postId, true);
+					this.attrs.goToPost(postId, true);
 				} else {
-					this.sendAction('goToPost', postId);
+					this.attrs.goToPost(postId, false);
 				}
 			},
