+++ front/scripts/main/components/post-reply.js
@@ -1,8 +1,6 @@
-import DiscussionUpvoteActionSendMixin from '../mixins/discussion-upvote-action-send';
 import DiscussionParsedContentMixin from '../mixins/discussion-parsed-content';
 
 export default Ember.Component.extend(
-	DiscussionUpvoteActionSendMixin,
 	DiscussionParsedContentMixin,
 	{
 		classNames: ['post-reply'],
