+++ front/scripts/main/controllers/discussion/forum.js
@@ -8,6 +8,7 @@ export default Ember.Controller.extend(DiscussionDeleteControllerMixin, Discussi
 	smartBannerVisible: Ember.computed.oneWay('application.smartBannerVisible'),
 	siteHeadPinned: Ember.computed.oneWay('application.siteHeadPinned'),
 
+	// TODO fix sorting
 	// Whether the sort component is currently visible
 	sortVisible: false,
 
@@ -53,6 +54,23 @@ export default Ember.Controller.extend(DiscussionDeleteControllerMixin, Discussi
 		 */
 		retry() {
 			this.get('target').send('retry');
+		},
+
+		// TODO
+		create(postData) {
+			this.get('target').send('create', postData);
+		},
+
+		goToPost(postId, openInNewTab = false) {
+			this.get('target').send('goToPost', postId, openInNewTab);
+		},
+
+		loadPage(pageNum) {
+			this.get('target').send('loadPage', pageNum);
+		},
+
+		upvote(post) {
+			this.get('target').send('upvote', post);
 		},
 
 		/**
