+++ front/scripts/main/mixins/discussion-upvote-action-send.js
@@ -19,7 +19,7 @@ export default Ember.Mixin.create({
 		 */
 		upvote(post) {
 			if (this.canUpvote()) {
-				this.sendAction('upvote', post);
+				this.attrs.upvote(post);
 			}
 		}
 	}
