+++ front/templates/main/components/post-reply.hbs
@@ -3,7 +3,7 @@
 		<div class="upvotes">
 			{{discussion-upvote-reply
 				post=post
-				upvote='upvote'
+				upvote=(action this.attrs.upvote)
 			}}
 		</div>
 		{{user-avatar
@@ -16,8 +16,8 @@
 			&bull; {{time-ago post.creationDate.epochSecond}}
 		</span>
 		{{discussion-delete
-			delete=this.attrs.deleteReply
-			undelete=this.attrs.undeleteReply
+			delete=(action this.attrs.deleteReply)
+			undelete=(action this.attrs.undeleteReply)
 			post=post
 		}}
 	</div>
