+++ front/templates/main/discussion/forum.hbs
@@ -1,66 +1,11 @@
-{{discussion-header
-	hideSortComponent='hideSortComponent'
-	showSortComponent='showSortComponent'
-	showContent=true
-	smartBannerVisible=smartBannerVisible
-	sortBy=sortBy
-	sortMessageKey=sortMessageKey
+{{discussion-forum
+	model=model
+	create=(action 'create')
+	deletePost=(action 'deletePost')
+	goToAllDiscussions=(action 'goToAllDiscussions')
+	goToPost=(action 'goToPost')
+	loadPage=(action 'loadPage')
+	retry=(action 'retry')
+	undeletePost=(action 'undeletePost')
+	upvote=(action 'upvote')
 }}
-{{discussion-error-dialog
-	isVisible=shouldShowErrorMessage
-	errorMessage=errorMessage
-}}
-
-<div class="site-body-discussion post-list-view">
-	{{discussion-hero-unit
-		contributors=model.contributors
-		name=model.name
-		numberOfPosts=model.totalPosts
-	}}
-	<div class="discussion-page row">
-		<div class="rail large-3 columns">
-			{{community-badge}}
-			{{discussion-back-button
-				href='/'
-				label='main.back-to-wikia-link'
-			}}
-		</div>
-		{{#if model.connectionError}}
-			{{discussion-connection-error
-				retry=(action 'retry')
-			}}
-		{{else if model.notFoundError}}
-			{{discussion-not-found-error
-				goToAllDiscussions=(action 'goToAllDiscussions')
-			}}
-		{{else}}
-			<div class="small-12 large-6 columns">
-				{{discussion-sort
-					setSortBy='setSortBy'
-					sortBy=sortBy
-					sortTypes=sortTypes
-					sortVisible=sortVisible
-				}}
-				{{forum-wrapper
-					create='create'
-					deletePost=(action 'deletePost')
-					goToPost='goToPost'
-					loadPage='loadPage'
-					setSortBy='setSortBy'
-					undeletePost=(action 'undeletePost')
-					upvote='upvote'
-					minorError=model.minorError
-					name=model.name
-					pageNum=model.pageNum
-					posts=model.posts
-					postsDisplayed=model.posts.length
-					totalPosts=model.totalPosts
-					siteHeadPinned=siteHeadPinned
-				}}
-			</div>
-		{{/if}}
-		<div class="rail large-3 columns">
-			{{discussion-app-promotion}}
-		</div>
-	</div>
-</div>
